+++ dask/array/core.py
@@ -11,6 +11,7 @@ from toolz.curried import (identity, pipe, partition, concat, unique, pluck,
         merge)
 import numpy as np
 from ..utils import deepmap
+from ..async import inline
 from .. import threaded, core
 
 
@@ -657,6 +658,6 @@ def get(dsk, keys, get=threaded.get, **kwargs):
     """
     fast_functions=kwargs.get('fast_functions',
                              set([operator.getitem, np.transpose]))
-    dsk2 = threaded.inline(dsk, fast_functions=fast_functions)
+    dsk2 = inline(dsk, fast_functions=fast_functions)
     dsk3 = core.cull(dsk2, list(core.flatten(keys)))
     return get(dsk3, keys, **kwargs)
